Criação das migrations

+++ database/migrations/2021_01_23_033005_create_official_documents_table.php
@@ -16,7 +16,7 @@ class CreateOfficialDocumentsTable extends Migration
         Schema::create('official_documents', function (Blueprint $table) {
             $table->id();
             $table->enum('type', ['CPF', 'CNPJ']);
-            $table->integer('number');
+            $table->bigInteger('number');
             $table->timestamps();
         });
     }
